+++ src/api_server/src/lib.rs
@@ -30,70 +30,9 @@ use mmds::data_store::Mmds;
 use parsed_request::ParsedRequest;
 use utils::eventfd::EventFd;
 use vmm::default_syscalls;
-use vmm::vmm_config::boot_source::BootSourceConfig;
-use vmm::vmm_config::drive::BlockDeviceConfig;
 use vmm::vmm_config::instance_info::InstanceInfo;
-use vmm::vmm_config::logger::LoggerConfig;
-use vmm::vmm_config::machine_config::VmConfig;
-use vmm::vmm_config::net::{NetworkInterfaceConfig, NetworkInterfaceUpdateConfig};
-use vmm::vmm_config::vsock::VsockDeviceConfig;
-use vmm::VmmActionError;
 
-/// This enum represents the public interface of the VMM. Each action contains various
-/// bits of information (ids, paths, etc.).
-#[derive(PartialEq)]
-pub enum VmmAction {
-    /// Configure the boot source of the microVM using as input the `ConfigureBootSource`. This
-    /// action can only be called before the microVM has booted.
-    ConfigureBootSource(BootSourceConfig),
-    /// Configure the logger using as input the `LoggerConfig`. This action can only be called
-    /// before the microVM has booted.
-    ConfigureLogger(LoggerConfig),
-    /// Get the configuration of the microVM.
-    GetVmConfiguration,
-    /// Flush the metrics. This action can only be called after the logger has been configured.
-    FlushMetrics,
-    /// Add a new block device or update one that already exists using the `BlockDeviceConfig` as
-    /// input. This action can only be called before the microVM has booted.
-    InsertBlockDevice(BlockDeviceConfig),
-    /// Add a new network interface config or update one that already exists using the
-    /// `NetworkInterfaceConfig` as input. This action can only be called before the microVM has
-    /// booted.
-    InsertNetworkDevice(NetworkInterfaceConfig),
-    /// Set the vsock device or update the one that already exists using the
-    /// `VsockDeviceConfig` as input. This action can only be called before the microVM has
-    /// booted.
-    SetVsockDevice(VsockDeviceConfig),
-    /// Update the size of an existing block device specified by an ID. The ID is the first data
-    /// associated with this enum variant. This action can only be called after the microVM is
-    /// started.
-    RescanBlockDevice(String),
-    /// Set the microVM configuration (memory & vcpu) using `VmConfig` as input. This
-    /// action can only be called before the microVM has booted.
-    SetVmConfiguration(VmConfig),
-    /// Launch the microVM. This action can only be called before the microVM has booted.
-    StartMicroVm,
-    /// Send CTRL+ALT+DEL to the microVM, using the i8042 keyboard function. If an AT-keyboard
-    /// driver is listening on the guest end, this can be used to shut down the microVM gracefully.
-    #[cfg(target_arch = "x86_64")]
-    SendCtrlAltDel,
-    /// Update the path of an existing block device. The data associated with this variant
-    /// represents the `drive_id` and the `path_on_host`.
-    UpdateBlockDevicePath(String, String),
-    /// Update a network interface, after microVM start. Currently, the only updatable properties
-    /// are the RX and TX rate limiters.
-    UpdateNetworkInterface(NetworkInterfaceUpdateConfig),
-}
-
-/// The enum represents the response sent by the VMM in case of success. The response is either
-/// empty, when no data needs to be sent, or an internal VMM structure.
-#[derive(Debug)]
-pub enum VmmData {
-    /// No data is sent on the channel.
-    Empty,
-    /// The microVM configuration represented by `VmConfig`.
-    MachineConfiguration(VmConfig),
-}
+use vmm::controller::{VmmAction, VmmRequest, VmmResponse};
 
 pub enum Error {
     Io(io::Error),
@@ -119,9 +58,6 @@ impl fmt::Debug for Error {
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
-
-pub type VmmRequest = Box<VmmAction>;
-pub type VmmResponse = Box<std::result::Result<VmmData, VmmActionError>>;
 
 pub struct ApiServer {
     /// MMDS info directly accessible from the API thread.
